requirements.txt
astroid==2.0.4
autopep8==1.4.3
certifi==2018.10.15
colorama==0.4.0
dj-database-url==0.5.0
Django==2.1.1
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pipenv==2018.10.13
psycopg2==2.7.5
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
six==1.11.0
virtualenv==16.0.0
virtualenv-clone==0.4.0
whitenoise==4.1
wrapt==1.10.11
